refactor(tx-record): TransactionRecord, RealmTransaction 클래스에 non-null 타입이 가능한 프로퍼티 수정
- RealmWalletBase 안쓰는 프로퍼티 삭제

+++ test/mock/transaction_mock.dart
@@ -15,7 +15,7 @@ class TransactionMock {
     String? memo,
     int amount = 1000000,
     int fee = 10000,
-    int vSize = 250,
+    double vSize = 250,
     DateTime? createdAt,
   }) {
     return TransactionRecord(
